+++ tests/components/enphase_ev/test_api_urls.py
@@ -25,6 +25,7 @@ async def test_api_builds_urls_correctly():
     await c.battery_status()
     await c.dry_contacts_settings()
     await c.grid_control_check()
+    await c.off_grid_due_to_grid_outage()
     await c.request_grid_toggle_otp()
     await c.validate_grid_toggle_otp("1234")
     await c.set_grid_state("ENV123", 1)
@@ -54,24 +55,28 @@ async def test_api_builds_urls_correctly():
     assert methods_urls[3][0] == "GET"
     assert f"/app-api/{RANDOM_SITE_ID}/grid_control_check.json" in methods_urls[3][1]
     assert methods_urls[4][0] == "GET"
-    assert f"/app-api/{RANDOM_SITE_ID}/grid_toggle_otp.json" in methods_urls[4][1]
-    assert methods_urls[5][0] == "POST"
-    assert "/app-api/grid_toggle_otp.json" in methods_urls[5][1]
-    assert methods_urls[6][0] == "POST"
-    assert "/pv/settings/grid_state.json" in methods_urls[6][1]
-    assert methods_urls[7][0] == "POST"
-    assert "/pv/settings/log_grid_change.json" in methods_urls[7][1]
-    assert methods_urls[8][0] == "GET"
     assert (
-        f"/app-api/{RANDOM_SITE_ID}/battery_backup_history.json" in methods_urls[8][1]
+        f"/app-api/{RANDOM_SITE_ID}/off_grid_due_to_grid_outage" in methods_urls[4][1]
     )
+    assert methods_urls[5][0] == "GET"
+    assert f"/app-api/{RANDOM_SITE_ID}/grid_toggle_otp.json" in methods_urls[5][1]
+    assert methods_urls[6][0] == "POST"
+    assert "/app-api/grid_toggle_otp.json" in methods_urls[6][1]
+    assert methods_urls[7][0] == "POST"
+    assert "/pv/settings/grid_state.json" in methods_urls[7][1]
+    assert methods_urls[8][0] == "POST"
+    assert "/pv/settings/log_grid_change.json" in methods_urls[8][1]
     assert methods_urls[9][0] == "GET"
-    assert f"/app-api/{RANDOM_SITE_ID}/inverters.json" in methods_urls[9][1]
-    assert methods_urls[10][0] == "GET"
-    assert f"/systems/{RANDOM_SITE_ID}/inverter_status_x.json" in methods_urls[10][1]
-    assert methods_urls[11][0] == "GET"
     assert (
-        f"/systems/{RANDOM_SITE_ID}/inverter_data_x/energy.json" in methods_urls[11][1]
+        f"/app-api/{RANDOM_SITE_ID}/battery_backup_history.json" in methods_urls[9][1]
+    )
+    assert methods_urls[10][0] == "GET"
+    assert f"/app-api/{RANDOM_SITE_ID}/inverters.json" in methods_urls[10][1]
+    assert methods_urls[11][0] == "GET"
+    assert f"/systems/{RANDOM_SITE_ID}/inverter_status_x.json" in methods_urls[11][1]
+    assert methods_urls[12][0] == "GET"
+    assert (
+        f"/systems/{RANDOM_SITE_ID}/inverter_data_x/energy.json" in methods_urls[12][1]
     )
     opt_out_call = next(
         (